transfer_learning/object_classification/mobnet/requirements.txt
tensorflow
tensorflow_hub
pandas
Pillow
opencv-python